move pointer reference classes into pointer utility
