+++ coilsnake/modules/eb/ExpandedTablesModule.py
@@ -2,45 +2,28 @@ import logging
 from coilsnake.model.eb.table import eb_table_from_offset
 from coilsnake.modules.eb.EbModule import EbModule
 from coilsnake.util.common.yml import yml_load
-from coilsnake.util.eb.pointer import write_asm_pointer, write_xl_pointer, from_snes_address, to_snes_address
-
-class AsmPointer(object):
-    def __init__(self, offset):
-        self.offset = offset
-
-    def write(self, rom, address):
-        log.info("Writing pointer at " + hex(self.offset))
-        write_asm_pointer(rom, self.offset, address)
-
-class XlPointer(object):
-    def __init__(self, offset):
-        self.offset = offset
-
-    def write(self, rom, address):
-        log.info("Writing xl pointer at " + hex(self.offset))
-        write_xl_pointer(rom, self.offset, address)
+from coilsnake.util.eb.pointer import from_snes_address, to_snes_address, AsmPointerReference, XlPointerReference
 
 log = logging.getLogger(__name__)
-
 
 
 class ExpandedTablesModule(EbModule):
     NAME = "Expanded Tables"
     TABLE_OFFSETS = {
         0xD58D7A: [  # PSI NAMES
-            AsmPointer(0x1c423)
+            AsmPointerReference(0x1c423)
         ],
         0xCF8985: [  # NPC Configuration Table
-            AsmPointer(0x023E5), # in func C0222B/ Load NPCs
-            XlPointer(0x0C32E), # in func C0C30C
-            AsmPointer(0x131CA), # in func C13187/Talk to
-            AsmPointer(0x1327F), # in func C1323B/Check NPC
-            AsmPointer(0x1332F), # in func C1323B/Check NPC
-            XlPointer(0x1AD77), # in func C1AD42
-            AsmPointer(0x1B20B), # in func C1AF74/Use overworld item
-            AsmPointer(0x464C1), # in func C464B5/Create prepared NPC
-            AsmPointer(0x46831), # in func C4681A
-            AsmPointer(0x46930), # in func C46914
+            AsmPointerReference(0x023E5), # in func C0222B/ Load NPCs
+            XlPointerReference(0x0C32E), # in func C0C30C
+            AsmPointerReference(0x131CA), # in func C13187/Talk to
+            AsmPointerReference(0x1327F), # in func C1323B/Check NPC
+            AsmPointerReference(0x1332F), # in func C1323B/Check NPC
+            XlPointerReference(0x1AD77), # in func C1AD42
+            AsmPointerReference(0x1B20B), # in func C1AF74/Use overworld item
+            AsmPointerReference(0x464C1), # in func C464B5/Create prepared NPC
+            AsmPointerReference(0x46831), # in func C4681A
+            AsmPointerReference(0x46930), # in func C46914
         ],
     }
